extract show_diffs to function
